set short format as default and add option to mark tasks as complete on a certain date; add yesterday and last/ as day descriptors

+++ src/format.rs
@@ -31,7 +31,7 @@ impl Iterator for TickTock {
 
 impl Default for TaskListFormatter {
     fn default() -> Self {
-        Self::Long
+        Self::Short
     }
 }
